+++ include/kiss3d-custom/src/resource/planar_mesh_manager.rs
@@ -59,11 +59,11 @@ impl PlanarMeshManager {
             circle_vtx.push(Point2::new(ang.cos(), ang.sin()) * 0.5);
             circle_ids.push(Point3::new(
                 0,
-                circle_vtx.len() as u16 - 2,
-                circle_vtx.len() as u16 - 1,
+                circle_vtx.len() as u32 - 2,
+                circle_vtx.len() as u32 - 1,
             ));
         }
-        circle_ids.push(Point3::new(0, circle_vtx.len() as u16 - 1, 1));
+        circle_ids.push(Point3::new(0, circle_vtx.len() as u32 - 1, 1));
 
         let circle = PlanarMesh::new(circle_vtx, circle_ids, None, false);
         res.add(Rc::new(RefCell::new(circle)), "circle");
