+++ include/kiss3d-custom/src/scene/planar_object.rs
@@ -229,7 +229,7 @@ impl PlanarObject {
 
     /// Mutably access the object's faces.
     #[inline(always)]
-    pub fn modify_faces<F: FnMut(&mut Vec<Point3<u16>>)>(&mut self, f: &mut F) {
+    pub fn modify_faces<F: FnMut(&mut Vec<Point3<u32>>)>(&mut self, f: &mut F) {
         let bmesh = self.mesh.borrow_mut();
         let _ = bmesh
             .faces()
@@ -242,7 +242,7 @@ impl PlanarObject {
 
     /// Access the object's faces.
     #[inline(always)]
-    pub fn read_faces<F: FnMut(&[Point3<u16>])>(&self, f: &mut F) {
+    pub fn read_faces<F: FnMut(&[Point3<u32>])>(&self, f: &mut F) {
         let bmesh = self.mesh.borrow();
         let _ = bmesh
             .faces()
